feat: move test code to sinotest mod.
