+++ SinoTest/sinotest-fabric/gradle.properties
@@ -1,0 +1,3 @@
+modid=sinotest
+mod_name=SinoTest
+mod_description=Test of SinoSeries.
