+++ SinoTest/sinotest-fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,31 @@
+{
+    "schemaVersion": 1,
+    "id": "${modid}",
+    "version": "${sinoseries_version}",
+    "name": "${mod_name}",
+    "description": "${mod_description}",
+    "authors": [
+        "${sinoseries_authors}"
+    ],
+    "contact": {
+        "homepage": "${sinoseries_url}",
+        "sources": "${sinoseries_source}",
+        "issues": "${sinoseries_issues}"
+    },
+    "license": "${sinoseries_license}",
+    "environment": "*",
+    "entrypoints": {
+        "main": [
+            "games.moegirl.sinocraft.sinotest.fabric.SinoTestFabric"
+        ]
+    },
+    "mixins": [
+        "sinotest-common.mixins.json"
+    ],
+    "depends": {
+        "java": ">=17",
+        "minecraft": "~${minecraft_version}",
+        "fabricloader": ">=${fabric_loader_version}",
+        "fabric-api": "*"
+    }
+}
