+++ SinoTest/sinotest-forge/gradle.properties
@@ -1,0 +1,5 @@
+loom.platform=forge
+
+modid=sinotest
+mod_name=SinoTest
+mod_description=Test of SinoSeries.
