+++ SinoTest/sinotest-forge/src/main/java/games/moegirl/sinocraft/sinotest/forge/SinoTestForge.java
@@ -1,0 +1,12 @@
+package games.moegirl.sinocraft.sinotest.forge;
+
+import games.moegirl.sinocraft.sinotest.SinoTest;
+import net.minecraftforge.fml.common.Mod;
+
+@Mod(SinoTest.MODID)
+public class SinoTestForge {
+
+    public SinoTestForge() {
+        SinoTest.registerAll();
+    }
+}
