+++ SinoTest/sinotest-forge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,32 @@
+modLoader = "javafml"
+loaderVersion = "[${fmlloader_version},${next_fmlloader_version})"
+license = "${sinoseries_license}"
+issueTrackerURL = "${sinoseries_issues}"
+
+[[mods]]
+modId = "${modid}"
+version = "${sinoseries_version}"
+displayName = "${mod_name}"
+#updateJSONURL="https://change.me.example.invalid/updates.json"
+displayURL = "${sinoseries_url}"
+#logoFile=""
+credits = "${sinoseries_credits}"
+authors = "${sinoseries_authors}"
+displayTest="MATCH_VERSION"
+description = '''
+${mod_description}
+'''
+
+[[dependencies.${modid}]]
+modId = "forge"
+mandatory = true
+versionRange = "[${forge_version},${next_forge_version})"
+ordering = "NONE"
+side = "BOTH"
+
+[[dependencies.${modid}]]
+modId = "minecraft"
+mandatory = true
+versionRange = "[${minecraft_version},${next_minecraft_version})"
+ordering = "NONE"
+side = "BOTH"
