+++ SinoTest/sinotest-neoforge/gradle.properties
@@ -1,0 +1,5 @@
+loom.platform=neoforge
+
+modid=sinotest
+mod_name=SinoTest
+mod_description=Test of SinoSeries.
