+++ SinoTest/sinotest-neoforge/src/main/java/games/moegirl/sinocraft/sinotest/neoforge/SinoTestNeoForge.java
@@ -1,0 +1,12 @@
+package games.moegirl.sinocraft.sinotest.neoforge;
+
+import games.moegirl.sinocraft.sinotest.SinoTest;
+import net.neoforged.fml.common.Mod;
+
+@Mod(SinoTest.MODID)
+public class SinoTestNeoForge {
+
+    public SinoTestNeoForge() {
+        SinoTest.registerAll();
+    }
+}
