+++ SinoTest/sinotest-neoforge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,35 @@
+modLoader = "javafml"
+loaderVersion = "[${neoforge_loader_version},${next_neoforge_loader_version})"
+license = "${sinoseries_license}"
+issueTrackerURL = "${sinoseries_issues}"
+
+[[mods]]
+modId = "${modid}"
+version = "${sinoseries_version}"
+displayName = "${mod_name}"
+#updateJSONURL="https://change.me.example.invalid/updates.json"
+displayURL = "${sinoseries_url}"
+#logoFile=""
+credits = "${sinoseries_credits}"
+authors = "${sinoseries_authors}"
+displayTest="MATCH_VERSION"
+description = '''
+${mod_description}
+'''
+
+[[mixins]]
+config = "sinocore-common.mixins.json"
+
+[[dependencies.${modid}]]
+modId = "neoforge"
+mandatory = true
+versionRange = "[${neoforge_version},${next_neoforge_version})"
+ordering = "NONE"
+side = "BOTH"
+
+[[dependencies.${modid}]]
+modId = "minecraft"
+mandatory = true
+versionRange = "[${minecraft_version},${next_minecraft_version})"
+ordering = "NONE"
+side = "BOTH"
