+++ SinoTest/src/main/java/games/moegirl/sinocraft/sinotest/registry/TestRegistryData.java
@@ -1,0 +1,4 @@
+package games.moegirl.sinocraft.sinotest.registry;
+
+public record TestRegistryData(String value) {
+}
