+++ SinoTest/src/main/resources/architectury.common.json
@@ -1,0 +1,4 @@
+{
+    "injected_interfaces": {
+    }
+}
