+++ SinoTest/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+    "pack": {
+        "description": "${modid}",
+        "pack_format": 15
+    }
+}
