+++ sinocore/sinocore-fabric/src/main/java/games/moegirl/sinocraft/sinocore/fabric/SinoCoreFabric.java
@@ -6,12 +6,11 @@ import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.minecraft.server.MinecraftServer;
 
 public class SinoCoreFabric implements ModInitializer {
-
     @Override
     public void onInitialize() {
         ServerLifecycleEvents.SERVER_STARTED.register(server -> SinoCorePlatformImpl.SERVER = server);
         ServerLifecycleEvents.SERVER_STOPPED.register(server -> SinoCorePlatformImpl.SERVER = null);
 
-        SinoCore.registerAll();
+        new SinoCore().init();
     }
 }
