+++ sinocore/sinocore-forge/src/main/java/games/moegirl/sinocraft/sinocore/forge/SinoCoreForge.java
@@ -5,8 +5,7 @@ import net.minecraftforge.fml.common.Mod;
 
 @Mod(SinoCore.MODID)
 public class SinoCoreForge {
-
     public SinoCoreForge() {
-        SinoCore.registerAll();
+        new SinoCore().init();
     }
 }
