+++ sinocore/sinocore-neoforge/src/main/java/games/moegirl/sinocraft/sinocore/neoforge/SinoCoreNeoForge.java
@@ -5,8 +5,7 @@ import net.neoforged.fml.common.Mod;
 
 @Mod(SinoCore.MODID)
 public class SinoCoreNeoForge {
-
     public SinoCoreNeoForge() {
-        SinoCore.registerAll();
+        new SinoCore().init();
     }
 }
