--- sinocore/src/main/java/games/moegirl/sinocraft/sinocore/test/registry/TestRegistryData.java
@@ -1,4 +1,0 @@
-package games.moegirl.sinocraft.sinocore.test.registry;
-
-public record TestRegistryData(String value) {
-}
